+++ src/features/archive/ui/ColorChip.module.scss
