+++ src/features/index.ts
@@ -1,1 +1,1 @@
-export * from './temp';
+export * from './archive';
